% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srlTS.R
\name{srlTS}
\alias{srlTS}
\alias{plot.srlTS}
\alias{coef.srlTS}
\alias{print.srlTS}
\alias{summary.srlTS}
\title{Perform time series ranked sparsity methods}
\usage{
srlTS(
  y,
  X = NULL,
  n_lags_max,
  gamma = c(0, 2^(-2:4)),
  ptrain = 0.8,
  pf_eps = 0.01,
  w_endo,
  w_exo,
  ncvreg_args = list(penalty = "lasso", returnX = FALSE, lambda.min = 0.001)
)

\method{plot}{srlTS}(x, log.l = TRUE, ...)

\method{coef}{srlTS}(object, choose = c("AICc", "BIC", "all"), ...)

\method{print}{srlTS}(x, ...)

\method{summary}{srlTS}(object, ...)
}
\arguments{
\item{y}{univariate time series outcome}

\item{X}{matrix of predictors (no intercept)}

\item{n_lags_max}{maximum number of lags to consider}

\item{gamma}{vector of exponent for weights}

\item{ptrain}{prop. to leave out for test data}

\item{pf_eps}{penalty factors below this will be set to zero}

\item{w_endo}{optional pre-specified weights for endogenous terms}

\item{w_exo}{optional pre-specified weights for exogenous terms (see details)}

\item{ncvreg_args}{additional args to pass through to ncvreg}

\item{x}{a srlTS object}

\item{log.l}{Should the x-axis (lambda) be logged?}

\item{...}{passed to downstream functions}

\item{object}{a srlTS object}

\item{choose}{which criterion to use for lambda selection (AICc, BIC, or all)}
}
\value{
A list of class \code{slrTS} with elements

  \item{fits}{a list of lasso fits}
  \item{ncvreg_args}{arguments passed to ncvreg}
  \item{gamma}{the (negative) exponent on the penalty weights, one
  for each fit}
  \item{n_lags_max}{the maximum number of lags}
  \item{y}{the time series}
  \item{X}{the utilized matrix of exogenous features}
  \item{oos_results}{results on test data using best of fits}
  \item{train_idx}{index of observations used in training data}

x invisibly

a vector of model coefficients

x (invisibly)

the summary object produced by ncvreg
  evaluated at the best tuning parameter combination
  (best AICc).
}
\description{
Perform time series ranked sparsity methods
}
\details{
The default weights for exogenous features will be chosen based on
a similar approach to the adaptive lasso (using bivariate OLS estimates). For
lower dimensional X, it's advised to set \code{w_exo="unpenalized"}, because
this allows for statistical inference on exogenous variable coefficients
via the \code{summary} function.
}
\examples{
data("LakeHuron")
fit_LH <- srlTS(LakeHuron)
fit_LH
coef(fit_LH)
plot(fit_LH)

}
\references{
Breheny, P. and Huang, J. (2011) Coordinate descent algorithms
  for nonconvex penalized regression, with applications to biological feature
  selection. Ann. Appl. Statist., 5: 232-253.

  Peterson, R.A., Cavanaugh, J.E. Ranked sparsity: a cogent regularization
  framework for selecting and estimating feature interactions and
  polynomials. AStA Adv Stat Anal (2022).
  https://doi.org/10.1007/s10182-021-00431-7
}
\seealso{
predict.srlTS
}
